ccl: cleanup pass — stale comments, dead file, warnings, sweep default
- device_backend: replace the outdated lazy-worker comment with the
  accurate two-attempt history (CUDA 13.3 first-launch hang + B300
  multi-block stall) and why pre-created workers are kept.
- alltoall_comparison: drop the stale 'lazy worker is live' text and the
  outdated tuning-levers note; point to the post-fix sweep.
- backend: delete the comment-only backend.cc (note moved to backend.h),
  drop it from the Makefile targets.
- executor: clamp the ready-ring size to a valid jring power-of-two and
  guard the jring size result (silences -Walloc-size-larger-than; the
  old path could compute SIZE_MAX for degenerate op counts).
- executor_factory: fix size_t->uint32_t smem_size narrowing warning.
- bench_shim_param_sweep: drop 128 blocks from the default DEV_BLOCKS
  sweep (unrealistic for the low-SM goal; still overridable via env).

+++ experimental/ukernel/src/ccl/executor_factory.cc
@@ -58,7 +58,7 @@ std::unique_ptr<SprayExecutor> SprayExecutor::create(
       .threads_per_block = static_cast<uint32_t>(config.threads_per_block),
       .blocks_per_worker = 0,  // resolved below (env / config / auto)
       .fifo_capacity = static_cast<uint32_t>(config.fifo_capacity),
-      .smem_size = config.smem_size,
+      .smem_size = static_cast<uint32_t>(config.smem_size),
       .idle_exit_after_us = config.device_idle_exit_us,
   };
   // Optional env overrides for benchmarking (win over config values):
